core: Don't create an endpoint per thread
Remove the functionality to create an endpoint per thread.  Originally,
this was used to create a full domain/endpoint/cq set of objects per
thread that created a communicator, but that got messy as the rdma
transport created multiple endpoints per thread based on the comm
in use.  So we ended up in a place where we were creating a domain
per thread (sometimes) and an endpoint per thread (always), which was
messy but worked.

With the switch from requesting FI_THREAD_SAFE to FI_THREAD_DOMAIN and
the concurrent switch to domain-level locking for plugin operations,
there really isn't much advantage to the endpoint per thread model,
so this patch removes all that logic.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>